move `MarkdownStoryWriter` tests to separate file
