+++ tests/shared/core/test_trackers.py
@@ -65,9 +65,6 @@ from tests.core.utilities import (
     get_tracker,
 )
 
-from rasa.shared.core.training_data.story_writer.markdown_story_writer import (
-    MarkdownStoryWriter,
-)
 from rasa.shared.nlu.constants import ACTION_NAME, PREDICTED_CONFIDENCE_KEY
 
 domain = Domain.load("examples/moodbot/domain.yml")
@@ -633,22 +630,6 @@ def test_session_started_not_part_of_applied_events(default_agent: Agent):
     # the SessionStart event was at index 5, the tracker's `applied_events()` should
     # be the same as the list of events from index 6 onwards
     assert tracker.applied_events() == list(tracker.events)[6:]
-
-
-async def test_tracker_dump_e2e_story(default_agent: Agent):
-    sender_id = "test_tracker_dump_e2e_story"
-
-    await default_agent.handle_text("/greet", sender_id=sender_id)
-    await default_agent.handle_text("/goodbye", sender_id=sender_id)
-    tracker = default_agent.tracker_store.get_or_create_tracker(sender_id)
-
-    story = tracker.export_stories(MarkdownStoryWriter(), e2e=True)
-    assert story.strip().split("\n") == [
-        "## test_tracker_dump_e2e_story",
-        "* greet: /greet",
-        "    - utter_greet",
-        "* goodbye: /goodbye",
-    ]
 
 
 def test_get_last_event_for():
